Stop exporting TH/THC symbols as extern "C".
This is back from the old days when we had C clients, but we no
longer have this.  Switch off extern "C".

This makes us less likely to run into a ROCm bug where functions
declared __host__ do not respect extern "C" in the header forward
declaration (you have to explicitly respecify it on the implementation.)

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 2391e6da858287f20a681e24d9324a72cd0594cb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29919